backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
